refactoring controllers and middlewares
