+++ src/middlewares/usertokenverifcation.js
@@ -1,6 +1,7 @@
 import jwt from 'jsonwebtoken';
 import 'dotenv/config';
 import VerifyTokenError from '../utils/verifytokenerror';
+import UserServices from '../services/user.service';
 // verify user token
 const verifyUserToken = async (req, res, next) => {
   try {
@@ -11,6 +12,9 @@ const verifyUserToken = async (req, res, next) => {
     if (!token) throw new VerifyTokenError('No token found', 401);
     const verify = jwt.verify(token, process.env.TOKEN_SECRET);
     if (!verify) throw new VerifyTokenError('token cant be verified', 401);
+    const decodedToken = jwt.decode(token);
+    const record = await UserServices.getUserById(decodedToken.data);
+    if (!record) throw new VerifyTokenError('data in token is invalid', 401);
     res.locals = token;
     if (res.locals !== token) throw new VerifyTokenError('server cant assign token', 500);
     next();
